+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Fixed
 
 - context: Fixed order of `add_contract` method arguments.
-- sentry: Fix sending crash reports when DSN is not set implicitly.
+- index: Fixed matching operations when both `address` and `code_hash` filters are specified.
+- sentry: Fixed sending crash reports when DSN is not set implicitly.
 - sentry: Increase event length limit.
 
 ## [6.4.2] - 2022-12-31